+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is used to list changes made in each version of the chatmail cookbook.
 
-## HEAD
+## 0.3.2
 
 ### Changes